Formated the Makefile and continued working on getting a player character on screen.

+++ include/player.h
@@ -3,5 +3,5 @@
 
 typedef struct player Player;
 
-Player *createPlayer(int x, int y, SDL_Renderer *pRenderer, int window_width, int window_height);
+Player *createPlayer(float x, float y, SDL_Renderer *pRenderer, int window_width, int window_height);
 #endif
